requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
checksumdir==1.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
filelock==3.13.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.47.0
fsspec==2024.3.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.22.2
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.8.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.6.0
networkx==3.3
numpy==1.26.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.14.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==5.28.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.38.0
sympy==1.13.1
tbb==2021.12.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
watchdog==4.0.2
Werkzeug==3.0.3
wrapt==1.17.2
